The MPLAT property editor no longer crashed, but also no longer renders.

+++ source/rococo/rococo.util/joysticks.xbox360.cpp
@@ -162,7 +162,7 @@ namespace
 					return GetVKeyCode(keyName);
 				}
 
-				SetLastError(E_INVALIDARG);
+				SetLastError((DWORD) E_INVALIDARG);
 				return 0;
 			}
